Update XPostBot.py
New timing system. Bot now decides how long to sleep based upon the time of the last submission it made is.

Changed path finding syntax to remove the dependency of the file needing to be named 'XPostBot.py'. Only restriction is to not have any spaces or non-alphanumeric characters in the filename.

Slight beautification of console output.